refactor: scrap operations and protocol
After an audit of caldav/operations/ and caldav/protocol/ (documented in
docs/design/OPERATIONS_PROTOCOL_AUDIT.md), both directories are deleted,
the code that was in use have been moved elsewhere.

The code changes are predominantly AI-written.  Tedious refactoring
work, chances for mistakes are bigger when doing it by hand than by AI.
I've been looking through the changes, and I trust the tests to uncover
any errors slipping through.

prompt: During the great attempt on Sans-IO refactoring, a directory
`caldav/operations/` was made. Please check up how much code is
duplicated and/or dead there, and come with recommendations on whether
to keep "operations" there or not. Same with the protocols folder.
followup-prompt: Save the analysis to the docs/design folder
followup-prompt: Kill the operations directory ref the document
followup-prompt: All response-related logic in the protocol directory
should be moved back to the response class. Make sure there is no
duplicated code or logic.
followup-prompt: move xml builders to the dav base client, and ensure
sync and async code paths uses the same builder methods
prompt: Deal with the code duplication in response.py
followup-prompt: It seems like the last commit, with purpose "remove code duplication in response.py" has more code additions than removed code?
