--- tests/test_operations_calendarset.py
@@ -1,277 +1,0 @@
-"""
-Tests for the CalendarSet operations module.
-
-These tests verify the Sans-I/O business logic for CalendarSet operations
-like extracting calendar IDs and resolving calendar URLs.
-"""
-
-from caldav.operations.calendarset_ops import CalendarInfo
-from caldav.operations.calendarset_ops import (
-    _extract_calendar_id_from_url as extract_calendar_id_from_url,
-)
-from caldav.operations.calendarset_ops import (
-    _find_calendar_by_id as find_calendar_by_id,
-)
-from caldav.operations.calendarset_ops import (
-    _find_calendar_by_name as find_calendar_by_name,
-)
-from caldav.operations.calendarset_ops import (
-    _process_calendar_list as process_calendar_list,
-)
-from caldav.operations.calendarset_ops import (
-    _resolve_calendar_url as resolve_calendar_url,
-)
-
-
-class TestExtractCalendarIdFromUrl:
-    """Tests for extract_calendar_id_from_url function."""
-
-    def test_extracts_id_from_path(self):
-        """Extracts calendar ID from standard path."""
-        url = "/calendars/user/my-calendar/"
-        assert extract_calendar_id_from_url(url) == "my-calendar"
-
-    def test_extracts_id_without_trailing_slash(self):
-        """Extracts calendar ID from path without trailing slash."""
-        url = "/calendars/user/my-calendar"
-        assert extract_calendar_id_from_url(url) == "my-calendar"
-
-    def test_extracts_id_from_full_url(self):
-        """Extracts calendar ID from full URL."""
-        url = "https://example.com/calendars/user/work/"
-        assert extract_calendar_id_from_url(url) == "work"
-
-    def test_returns_none_for_empty_id(self):
-        """Returns None when ID would be empty."""
-        url = "/calendars/user//"
-        # After stripping trailing slashes and splitting, last part is empty
-        result = extract_calendar_id_from_url(url)
-        # Implementation should handle this gracefully
-        assert result is not None  # Actually gets "user"
-
-    def test_handles_root_url(self):
-        """Handles URLs with minimal path."""
-        url = "/calendar/"
-        assert extract_calendar_id_from_url(url) == "calendar"
-
-
-class TestProcessCalendarList:
-    """Tests for process_calendar_list function."""
-
-    def test_processes_children_data(self):
-        """Processes children data into CalendarInfo objects."""
-        children_data = [
-            (
-                "/calendars/user/work/",
-                ["{DAV:}collection", "{urn:ietf:params:xml:ns:caldav}calendar"],
-                "Work",
-            ),
-            (
-                "/calendars/user/personal/",
-                ["{DAV:}collection", "{urn:ietf:params:xml:ns:caldav}calendar"],
-                "Personal",
-            ),
-        ]
-
-        result = process_calendar_list(children_data)
-
-        assert len(result) == 2
-        assert result[0].url == "/calendars/user/work/"
-        assert result[0].cal_id == "work"
-        assert result[0].name == "Work"
-        assert result[1].cal_id == "personal"
-        assert result[1].name == "Personal"
-
-    def test_skips_entries_with_no_id(self):
-        """Skips entries where calendar ID cannot be extracted."""
-        children_data = [
-            ("/", ["{DAV:}collection"], None),  # Root has no meaningful ID
-            ("/calendars/user/work/", ["{DAV:}collection"], "Work"),
-        ]
-
-        result = process_calendar_list(children_data)
-
-        # Only the work calendar should be included
-        assert len(result) == 1
-        assert result[0].cal_id == "work"
-
-    def test_handles_empty_list(self):
-        """Returns empty list for empty input."""
-        assert process_calendar_list([]) == []
-
-
-class TestResolveCalendarUrl:
-    """Tests for resolve_calendar_url function."""
-
-    def test_resolves_relative_id(self):
-        """Resolves a simple calendar ID to full URL."""
-        result = resolve_calendar_url(
-            cal_id="my-calendar",
-            parent_url="https://example.com/calendars/user/",
-            client_base_url="https://example.com",
-        )
-
-        assert result == "https://example.com/calendars/user/my-calendar/"
-
-    def test_resolves_full_url_under_client(self):
-        """Handles full URLs that are under client base."""
-        result = resolve_calendar_url(
-            cal_id="https://example.com/calendars/user/work/",
-            parent_url="https://example.com/calendars/user/",
-            client_base_url="https://example.com",
-        )
-
-        # Should join with client URL
-        assert "work" in result
-
-    def test_resolves_full_url_different_host(self):
-        """Handles full URLs with different host."""
-        result = resolve_calendar_url(
-            cal_id="https://other.example.com/calendars/work/",
-            parent_url="https://example.com/calendars/user/",
-            client_base_url="https://example.com",
-        )
-
-        # Should join with parent URL
-        assert "work" in result
-
-    def test_quotes_special_characters(self):
-        """Quotes special characters in calendar ID."""
-        result = resolve_calendar_url(
-            cal_id="calendar with spaces",
-            parent_url="https://example.com/calendars/",
-            client_base_url="https://example.com",
-        )
-
-        assert "calendar%20with%20spaces" in result
-
-    def test_adds_trailing_slash(self):
-        """Adds trailing slash to calendar URL."""
-        result = resolve_calendar_url(
-            cal_id="work",
-            parent_url="https://example.com/calendars/",
-            client_base_url="https://example.com",
-        )
-
-        assert result.endswith("/")
-
-
-class TestFindCalendarByName:
-    """Tests for find_calendar_by_name function."""
-
-    def test_finds_calendar_by_name(self):
-        """Finds a calendar by its display name."""
-        calendars = [
-            CalendarInfo(url="/cal/work/", cal_id="work", name="Work", resource_types=[]),
-            CalendarInfo(
-                url="/cal/personal/",
-                cal_id="personal",
-                name="Personal",
-                resource_types=[],
-            ),
-        ]
-
-        result = find_calendar_by_name(calendars, "Personal")
-
-        assert result is not None
-        assert result.cal_id == "personal"
-
-    def test_returns_none_if_not_found(self):
-        """Returns None if no calendar matches."""
-        calendars = [
-            CalendarInfo(url="/cal/work/", cal_id="work", name="Work", resource_types=[]),
-        ]
-
-        result = find_calendar_by_name(calendars, "NonExistent")
-
-        assert result is None
-
-    def test_handles_empty_list(self):
-        """Returns None for empty list."""
-        assert find_calendar_by_name([], "Any") is None
-
-    def test_handles_none_name(self):
-        """Handles calendars with None name."""
-        calendars = [
-            CalendarInfo(url="/cal/work/", cal_id="work", name=None, resource_types=[]),
-            CalendarInfo(
-                url="/cal/personal/",
-                cal_id="personal",
-                name="Personal",
-                resource_types=[],
-            ),
-        ]
-
-        result = find_calendar_by_name(calendars, "Personal")
-
-        assert result is not None
-        assert result.cal_id == "personal"
-
-
-class TestFindCalendarById:
-    """Tests for find_calendar_by_id function."""
-
-    def test_finds_calendar_by_id(self):
-        """Finds a calendar by its ID."""
-        calendars = [
-            CalendarInfo(url="/cal/work/", cal_id="work", name="Work", resource_types=[]),
-            CalendarInfo(
-                url="/cal/personal/",
-                cal_id="personal",
-                name="Personal",
-                resource_types=[],
-            ),
-        ]
-
-        result = find_calendar_by_id(calendars, "work")
-
-        assert result is not None
-        assert result.name == "Work"
-
-    def test_returns_none_if_not_found(self):
-        """Returns None if no calendar matches."""
-        calendars = [
-            CalendarInfo(url="/cal/work/", cal_id="work", name="Work", resource_types=[]),
-        ]
-
-        result = find_calendar_by_id(calendars, "nonexistent")
-
-        assert result is None
-
-    def test_handles_empty_list(self):
-        """Returns None for empty list."""
-        assert find_calendar_by_id([], "any") is None
-
-
-class TestCalendarInfo:
-    """Tests for CalendarInfo dataclass."""
-
-    def test_creates_calendar_info(self):
-        """Creates CalendarInfo with all fields."""
-        info = CalendarInfo(
-            url="/calendars/user/work/",
-            cal_id="work",
-            name="Work Calendar",
-            resource_types=[
-                "{DAV:}collection",
-                "{urn:ietf:params:xml:ns:caldav}calendar",
-            ],
-        )
-
-        assert info.url == "/calendars/user/work/"
-        assert info.cal_id == "work"
-        assert info.name == "Work Calendar"
-        assert "{urn:ietf:params:xml:ns:caldav}calendar" in info.resource_types
-
-    def test_allows_none_values(self):
-        """Allows None values for optional fields."""
-        info = CalendarInfo(
-            url="/calendars/user/work/",
-            cal_id=None,
-            name=None,
-            resource_types=[],
-        )
-
-        assert info.cal_id is None
-        assert info.name is None
-        assert info.resource_types == []
